refactor: update user model to use 'username' instead of 'name'

+++ app/api/info/route.js
@@ -1,25 +1,61 @@
 
 import { connectMongoDB } from "@/lib/mongodb";
 import User from "@/models/user";
+import { getServerSession } from "next-auth";
 import { NextResponse } from "next/server";
+import { authOptions } from "@/lib/auth";
 
 
-export async function POST(request) {
-  const { name, email, password, salt } = await request.json();
+export async function POST() {
+  const session = await getServerSession(authOptions);
+
+  if (!session) {
+    return NextResponse.json({ message: "Unauthorized" }, { status: 401 });
+  }
+
   await connectMongoDB();
-  await User.create({ name, email, password, salt });
+  await User.create({ username, email, password, salt });
+
+  if (!user) {
+    return NextResponse.json({ message: "User not found" }, { status: 404 });
+  }
+
   return NextResponse.json({ message: "User Created" }, { status: 201 });
 }
 
+
 export async function GET() {
+  const session = await getServerSession(authOptions);
+  
+  if (!session) {
+    return NextResponse.json({ message: "Unauthorized" }, { status: 401 });
+  }
+
   await connectMongoDB();
-  const Details = await User.find();
-  return NextResponse.json({ Details }, { status: 200 });
+  const user = await User.findOne({ email: session.user.email });
+  
+  if (!user) {
+    return NextResponse.json({ message: "User not found" }, { status: 404 });
+  }
+
+  return NextResponse.json({Details: user}, { status: 200 });
 }
 
 export async function DELETE(request) {
+  const session = await getServerSession(authOptions);
+
+  if (!session) {
+    return NextResponse.json({ message: "Unauthorized" }, { status: 401 });
+  }
+
+    await connectMongoDB();
     const id = new URL(request.url).searchParams.get("id");
     await connectMongoDB();
     await User.findByIdAndDelete(id);
+
+    if (!user) {
+      return NextResponse.json({ message: "User not found" }, { status: 404 });
+    }
+
     return NextResponse.json({ message: "User deleted" }, { status: 200 });
 }
